requirements.txt
Flask==2.3.3
reportlab==4.0.4
Jinja2==3.1.2
Werkzeug==2.3.7
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
gunicorn==21.2.0